requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
google==3.0.0
google-api-core==2.24.0
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.0
google-cloud-secret-manager==2.22.1
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.71.0
grpcio-status==1.69.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
proto-plus==1.25.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-http-client==3.3.7
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.3
rsa==4.9.1
sendgrid==6.11.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
starkbank-ecdsa==2.2.0
starlette==0.46.2
starlette-authlib==0.3.10
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
google-cloud-storage

